refactor(cli): extract setup wizard into internal/setupwizard package
Addresses the GitHub CodeRabbit review on PR #236:

- (Major, repo guideline) New *_test.go files must live under tests/, not
  internal/. Moved the wizard's pure + IO logic (corpus profiles, dotenv
  detection, gitignore, key persistence, form build/run) into a dedicated
  internal/setupwizard package with an exported API, deleted
  internal/cli/setup_wizard{,_test}.go, and added coverage under
  tests/setupwizard/. The cli package keeps only command glue
  (eligibility, output, provider verification, git protection) and calls
  setupwizard.*. PersistKeys takes an injected dotenv writer so the cli
  retains control of file location/permissions (saveEnvLocalKey).
- (Major) Code corpus profile now sets RAGMaxContextChars explicitly so
  the preset is fully self-contained, matching legal.
- (Minor) README config init row now lists --quiet in the
  non-interactive/baseline path.

Removed both deleted files from the internal/cli file-ownership
allowlist. make check passes.
