+++ internal/setupwizard/wizard.go
@@ -1,4 +1,10 @@
-package cli
+// Package setupwizard holds the provider-credential and corpus-profile logic
+// behind `dir2mcp config init` and the `dir2mcp up` first-run flow (SPEC §§
+// config init / charmbracelet/huh). It is a standalone package so the pure
+// mapping and IO helpers are unit-testable from tests/ (the repo forbids new
+// *_test.go files under internal/); the CLI package wires these into command
+// output, provider verification, and persistence.
+package setupwizard
 
 import (
 	"errors"
@@ -11,22 +17,25 @@ import (
 	"github.com/dirstral/dir2mcp/internal/config"
 )
 
-// providerKeySpec describes one credential the setup wizard can collect. The
+// ProviderKeySpec describes one credential the setup wizard can collect. The
 // EnvVar is the dotenv key written to .env.local; secrets are never persisted
 // to the .dir2mcp.yaml snapshot (SPEC §16: api keys are env-sourced).
-type providerKeySpec struct {
+type ProviderKeySpec struct {
 	EnvVar      string
 	Title       string
 	Description string
 	Optional    bool // shown only when the user opts into optional providers
 }
 
-// wizardProviderKeys is the ordered set of credentials the wizard offers. The
-// required Mistral key and the rerank-enabling Cohere key are always shown; the
-// rest are revealed only when the user asks to configure optional providers, so
-// the common path stays a two-field form.
-var wizardProviderKeys = []providerKeySpec{
-	{EnvVar: "MISTRAL_API_KEY", Title: "Mistral API key", Description: "Required — embeddings, PDF/OCR extraction, and answers."},
+// MistralEnvVar is the required embedding/OCR/answer credential.
+const MistralEnvVar = "MISTRAL_API_KEY"
+
+// ProviderKeys is the ordered set of credentials the wizard offers. The required
+// Mistral key and the rerank-enabling Cohere key are always shown; the rest are
+// revealed only when the user asks to configure optional providers, so the
+// common path stays a two-field form.
+var ProviderKeys = []ProviderKeySpec{
+	{EnvVar: MistralEnvVar, Title: "Mistral API key", Description: "Required — embeddings, PDF/OCR extraction, and answers."},
 	{EnvVar: "COHERE_API_KEY", Title: "Cohere API key", Description: "Recommended — enables reranking (sharper citations)."},
 	{EnvVar: "OPENAI_API_KEY", Title: "OpenAI API key", Description: "Optional — alternate chat/embedding provider.", Optional: true},
 	{EnvVar: "ANTHROPIC_API_KEY", Title: "Anthropic API key", Description: "Optional — alternate chat provider.", Optional: true},
@@ -34,18 +43,16 @@ var wizardProviderKeys = []providerKeySpec{
 	{EnvVar: "ELEVENLABS_API_KEY", Title: "ElevenLabs API key", Description: "Optional — speech-to-text / text-to-speech.", Optional: true},
 }
 
-const mistralEnvVar = "MISTRAL_API_KEY"
-
-// corpusProfile is a named retrieval preset the wizard can apply to the config.
-type corpusProfile string
+// Profile is a named retrieval preset the wizard can apply to the config.
+type Profile string
 
 const (
-	// corpusProfileKeep leaves retrieval settings untouched (offered only when a
+	// ProfileKeep leaves retrieval settings untouched (offered only when a
 	// config already exists, so re-running the wizard need not re-tune).
-	corpusProfileKeep    corpusProfile = "keep"
-	corpusProfileGeneral corpusProfile = "general"
-	corpusProfileLegal   corpusProfile = "legal"
-	corpusProfileCode    corpusProfile = "code"
+	ProfileKeep    Profile = "keep"
+	ProfileGeneral Profile = "general"
+	ProfileLegal   Profile = "legal"
+	ProfileCode    Profile = "code"
 )
 
 const legalSystemPrompt = `You answer questions strictly from the provided legal documents: statutes,
@@ -60,7 +67,7 @@ documentation. Cite file paths and line ranges, and quote the relevant code.
 If the indexed code does not cover the question, say so plainly rather than
 guessing.`
 
-// applyCorpusProfile mutates cfg's retrieval settings to match the chosen
+// ApplyCorpusProfile mutates cfg's retrieval settings to match the chosen
 // profile. It is a pure function over cfg (no IO) so the mapping is unit
 // testable without driving the TUI.
 //
@@ -70,8 +77,8 @@ guessing.`
 // inherit a stale value tuned for the previous profile. "general" is therefore
 // the baseline (Config defaults). "keep" is the sole exception: it leaves cfg
 // untouched so an existing tuned config survives a re-run.
-func applyCorpusProfile(cfg *config.Config, profile corpusProfile) {
-	if profile == corpusProfileKeep {
+func ApplyCorpusProfile(cfg *config.Config, profile Profile) {
+	if profile == ProfileKeep {
 		return
 	}
 
@@ -81,60 +88,63 @@ func applyCorpusProfile(cfg *config.Config, profile corpusProfile) {
 	cfg.RAGSystemPrompt = def.RAGSystemPrompt
 
 	switch profile {
-	case corpusProfileLegal:
+	case ProfileLegal:
 		cfg.RAGKDefault = 12
 		cfg.RAGMaxContextChars = 40000
 		cfg.RAGSystemPrompt = legalSystemPrompt
-	case corpusProfileCode:
+	case ProfileCode:
+		// Code uses a tighter top-k and the standard context window; all three
+		// managed fields are set explicitly so the preset is self-contained.
 		cfg.RAGKDefault = 8
+		cfg.RAGMaxContextChars = def.RAGMaxContextChars
 		cfg.RAGSystemPrompt = codeSystemPrompt
 	default:
 		// general: Config defaults (already applied above).
 	}
 }
 
-// wizardResult holds the user's answers from the setup form: the collected
+// Result holds the user's answers from the setup form: the collected
 // credentials (keyed by env var, empty entries dropped) and the chosen corpus
 // profile.
-type wizardResult struct {
+type Result struct {
 	Keys    map[string]string
-	Profile corpusProfile
+	Profile Profile
 }
 
-// wizardInput parameterizes the setup form with what is already known about the
+// Input parameterizes the setup form with what is already known about the
 // environment so the form can adapt: which credentials are already set (so the
 // field can say "leave blank to keep" and the required check can pass), and
 // whether a config already exists (so a "keep current settings" profile option
 // is offered and pre-selected).
-type wizardInput struct {
+type Input struct {
 	ExistingKeys  map[string]bool
 	ConfigExisted bool
 }
 
 // keyDescription augments a provider field's static description with a note when
 // the credential is already present in the environment / .env.local.
-func keyDescription(spec providerKeySpec, existing map[string]bool) string {
+func keyDescription(spec ProviderKeySpec, existing map[string]bool) string {
 	if existing[spec.EnvVar] {
 		return spec.Description + " (already set — leave blank to keep)"
 	}
 	return spec.Description
 }
 
-// buildSetupForm constructs the huh form, binding each field to the provided
+// BuildForm constructs the huh form, binding each field to the provided
 // pointers. Optional provider inputs live in a group hidden until the user
-// confirms they want to configure more providers. Kept separate from execution
-// so the field/grouping layout can be reasoned about (and the form built) apart
-// from the interactive Run.
-func buildSetupForm(
+// confirms they want to configure more providers. Kept separate from Run so the
+// field/grouping layout can be reasoned about (and the form built) apart from
+// the interactive execution.
+func BuildForm(
 	keyValues map[string]*string,
 	configureMore *bool,
 	profile *string,
 	save *bool,
-	in wizardInput,
+	in Input,
 ) *huh.Form {
 	required := []huh.Field{}
 	optional := []huh.Field{}
-	for _, spec := range wizardProviderKeys {
+	for _, spec := range ProviderKeys {
 		input := huh.NewInput().
 			Title(spec.Title).
 			Description(keyDescription(spec, in.ExistingKeys)).
@@ -142,7 +152,7 @@ func buildSetupForm(
 			Value(keyValues[spec.EnvVar])
 		// Mistral is required: reject an empty value unless it is already set
 		// (in which case a blank field means "keep the existing key").
-		if spec.EnvVar == mistralEnvVar && !in.ExistingKeys[mistralEnvVar] {
+		if spec.EnvVar == MistralEnvVar && !in.ExistingKeys[MistralEnvVar] {
 			input = input.Validate(func(s string) error {
 				if strings.TrimSpace(s) == "" {
 					return errors.New("required — paste your Mistral API key (or pre-set MISTRAL_API_KEY)")
@@ -170,12 +180,12 @@ func buildSetupForm(
 
 	options := []huh.Option[string]{}
 	if in.ConfigExisted {
-		options = append(options, huh.NewOption("Keep current settings (don't change retrieval)", string(corpusProfileKeep)))
+		options = append(options, huh.NewOption("Keep current settings (don't change retrieval)", string(ProfileKeep)))
 	}
 	options = append(options,
-		huh.NewOption("General documents", string(corpusProfileGeneral)),
-		huh.NewOption("Legal / citations (strict grounding)", string(corpusProfileLegal)),
-		huh.NewOption("Source code", string(corpusProfileCode)),
+		huh.NewOption("General documents", string(ProfileGeneral)),
+		huh.NewOption("Legal / citations (strict grounding)", string(ProfileLegal)),
+		huh.NewOption("Source code", string(ProfileCode)),
 	)
 	profileGroup := huh.NewGroup(
 		huh.NewSelect[string]().
@@ -196,28 +206,28 @@ func buildSetupForm(
 	return huh.NewForm(providerGroup, optionalGroup, profileGroup, confirmGroup)
 }
 
-// runSetupWizard runs the interactive huh setup form and returns the user's
-// answers. Credentials are returned (not yet written) so the caller controls
-// persistence (.env.local) and config patching. Declining the final "Save"
-// confirm surfaces as huh.ErrUserAborted so callers treat it like Ctrl-C.
-func runSetupWizard(in wizardInput) (wizardResult, error) {
-	keyValues := make(map[string]*string, len(wizardProviderKeys))
-	for _, spec := range wizardProviderKeys {
+// Run runs the interactive huh setup form and returns the user's answers.
+// Credentials are returned (not yet written) so the caller controls persistence
+// (.env.local) and config patching. Declining the final "Save" confirm surfaces
+// as huh.ErrUserAborted so callers treat it like Ctrl-C.
+func Run(in Input) (Result, error) {
+	keyValues := make(map[string]*string, len(ProviderKeys))
+	for _, spec := range ProviderKeys {
 		keyValues[spec.EnvVar] = new(string)
 	}
 	var configureMore bool
 	save := true
-	profile := string(corpusProfileGeneral)
+	profile := string(ProfileGeneral)
 	if in.ConfigExisted {
-		profile = string(corpusProfileKeep)
+		profile = string(ProfileKeep)
 	}
 
-	form := buildSetupForm(keyValues, &configureMore, &profile, &save, in)
+	form := BuildForm(keyValues, &configureMore, &profile, &save, in)
 	if err := form.Run(); err != nil {
-		return wizardResult{}, err
+		return Result{}, err
 	}
 	if !save {
-		return wizardResult{}, huh.ErrUserAborted
+		return Result{}, huh.ErrUserAborted
 	}
 
 	keys := make(map[string]string)
@@ -226,30 +236,49 @@ func runSetupWizard(in wizardInput) (wizardResult, error) {
 			keys[env] = v
 		}
 	}
-	return wizardResult{Keys: keys, Profile: corpusProfile(profile)}, nil
+	return Result{Keys: keys, Profile: Profile(profile)}, nil
 }
 
-// detectExistingKeys reports which wizard credentials already resolve, either
+// PersistKeys writes each non-empty collected credential via write (a dotenv
+// upsert), iterating in ProviderKeys order for deterministic writes, and returns
+// the env-var names that were saved. The writer is injected so persistence is
+// testable and so the CLI controls file location/permissions.
+func PersistKeys(envPath string, keys map[string]string, write func(path, keyName, value string) error) ([]string, error) {
+	saved := make([]string, 0, len(keys))
+	for _, spec := range ProviderKeys {
+		v, ok := keys[spec.EnvVar]
+		if !ok || strings.TrimSpace(v) == "" {
+			continue
+		}
+		if err := write(envPath, spec.EnvVar, strings.TrimSpace(v)); err != nil {
+			return saved, err
+		}
+		saved = append(saved, spec.EnvVar)
+	}
+	return saved, nil
+}
+
+// DetectExistingKeys reports which wizard credentials already resolve, either
 // from the process environment or from a non-empty assignment in .env.local at
 // envPath. Used to relax the required check and annotate fields.
-func detectExistingKeys(envPath string) map[string]bool {
-	out := make(map[string]bool, len(wizardProviderKeys))
+func DetectExistingKeys(envPath string) map[string]bool {
+	out := make(map[string]bool, len(ProviderKeys))
 	var dotenv string
 	if b, err := os.ReadFile(envPath); err == nil {
 		dotenv = string(b)
 	}
-	for _, spec := range wizardProviderKeys {
-		if strings.TrimSpace(os.Getenv(spec.EnvVar)) != "" || dotenvHasKey(dotenv, spec.EnvVar) {
+	for _, spec := range ProviderKeys {
+		if strings.TrimSpace(os.Getenv(spec.EnvVar)) != "" || DotenvHasKey(dotenv, spec.EnvVar) {
 			out[spec.EnvVar] = true
 		}
 	}
 	return out
 }
 
-// dotenvHasKey reports whether content has a non-empty assignment for key,
+// DotenvHasKey reports whether content has a non-empty assignment for key,
 // honoring an optional leading "export " and treating quoted-empty values
 // (KEY="" / KEY=”) as unset.
-func dotenvHasKey(content, key string) bool {
+func DotenvHasKey(content, key string) bool {
 	for _, line := range strings.Split(content, "\n") {
 		line = strings.TrimSpace(line)
 		line = strings.TrimPrefix(line, "export ")
@@ -274,10 +303,10 @@ func unquoteDotenvValue(val string) string {
 	return val
 }
 
-// ensureGitignoreEntries appends any missing entries to dir/.gitignore (creating
+// EnsureGitignoreEntries appends any missing entries to dir/.gitignore (creating
 // it if needed), so credentials and local state are not accidentally committed.
 // Existing entries and file content are preserved.
-func ensureGitignoreEntries(dir string, entries ...string) error {
+func EnsureGitignoreEntries(dir string, entries ...string) error {
 	path := filepath.Join(dir, ".gitignore")
 	var existing string
 	if b, err := os.ReadFile(path); err == nil {
